Componentize Location chip in preparation of de-chipping said chip
