+++ src/components/details/MoreLikeThis.tsx
@@ -1,16 +1,14 @@
 import React, { FC, useState } from 'react'
-import { Route, useHistory } from 'react-router-dom'
+import { Route } from 'react-router-dom'
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles'
-import { Popover } from '@material-ui/core'
-import { BiMapPin } from 'react-icons/bi'
 import { IoIosPeople } from 'react-icons/io'
 
 import { routes } from 'components/config/api'
 import { SwatchOnly } from 'components/legend'
-import { Chip, NeighborhoodList } from 'components/details'
-import { DialogCloseBtn } from 'components/generic/modals'
+import { Chip } from 'components/details'
 
 import * as Types from './types'
+import { LocationLink } from './LocationLink'
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -24,19 +22,13 @@ const useStyles = makeStyles((theme: Theme) =>
         marginLeft: '0.35rem',
       },
     },
-    popover: {
-      maxWidth: 350,
-      minWidth: 325,
-      padding: '1rem',
-    },
   })
 )
 
-export const MoreLikeThis: FC<Types.MoreLikeThisProps> = (props) => {
-  const { data, children, isInstance } = props
+export const MoreLikeThis: FC<Types.TonsOfData> = (props) => {
+  const { data, children } = props
   const classes = useStyles()
   const [anchorEl, setAnchorEl] = useState<HTMLDivElement | null>(null)
-  const history = useHistory()
 
   const {
     'World Region': WorldRegion,
@@ -44,70 +36,17 @@ export const MoreLikeThis: FC<Types.MoreLikeThisProps> = (props) => {
     countryImg,
     Macrocommunity: macro,
     worldRegionColor,
-    Town,
-    Neighborhood,
   } = data
-
-  const primaryLoc = Town || Neighborhood
-
-  const handleClose = () => setAnchorEl(null)
-
-  const handleClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    setAnchorEl(event.currentTarget)
-  }
-
-  const open = Boolean(anchorEl)
-  const id = open ? 'neighbs-menu-popover' : undefined
-
-  let LocationChip
-  const addlCount = isInstance && data['Additional Neighborhoods']?.length
-  const neighbsChipText = `${primaryLoc}${addlCount ? ` +${addlCount}` : ''}`
-  const explorePath = Town ? 'Town' : 'Neighborhood'
-
-  if (addlCount) {
-    LocationChip = (
-      <>
-        <Popover
-          id={id}
-          open={open}
-          anchorEl={anchorEl}
-          onClose={handleClose}
-          PaperProps={{ className: classes.popover, elevation: 12 }}
-          anchorOrigin={{ vertical: 'top', horizontal: 'center' }}
-          transformOrigin={{ vertical: 'center', horizontal: 'left' }}
-        >
-          <NeighborhoodList data={data} isInstance={isInstance} />
-          <DialogCloseBtn
-            tooltip="Close census menu"
-            onClose={() => handleClose()}
-          />
-        </Popover>
-        <Chip
-          text={neighbsChipText}
-          icon={<BiMapPin />}
-          title="Show neighborhood or town options"
-          handleClick={handleClick}
-        />
-      </>
-    )
-  } else {
-    LocationChip = (
-      <Chip
-        text={neighbsChipText}
-        icon={<BiMapPin />}
-        title={`View more languages spoken in ${primaryLoc}`}
-        handleClick={() =>
-          history.push(`/Explore/${explorePath}/${primaryLoc}`)
-        }
-      />
-    )
-  }
 
   // TODO: use TS on all mid-route paths, e.g. "World Region"
   return (
     <div className={classes.root}>
       <Route path="/Explore/Language/:language/:id" exact>
-        {LocationChip}
+        <LocationLink
+          anchorEl={anchorEl}
+          setAnchorEl={setAnchorEl}
+          data={data}
+        />
       </Route>
       <Chip
         text={WorldRegion}
